add create day2 policy methods in RestClientVraNgPrimitive
